nrf_security: Prefer size_t instead of int in Cracen
There were many cases that the driver had int statements
which compared int with size_t. This throws a warning in
some compilers (gcc: -Wsign-compare).

Change all the variables to size_t for consistency.

No functional change.

+++ subsys/nrf_security/src/drivers/cracen/cracenpsa/src/cracen_psa_jpake.c
@@ -145,7 +145,7 @@ static psa_status_t cracen_ecjpake_get_public_key(sx_pk_req *req,
 						  const uint8_t *secret)
 {
 	struct sx_pk_inops_ecp_mult inputs;
-	int opsz;
+	size_t opsz;
 	int status;
 
 	sx_pk_set_cmd(req, SX_PK_CMD_ECC_PTMUL);
